Add .halcyon directory with build config
This enables building the whole project with just `halcyon install` once
https://halcyon.sh/ has been set up.

The constraints were generated with the command `halcyon constraints`,
and AFAIK `halcyon constraints --ignore-all-constraints` should be used
to update them one day.

* * *

Something also messes up the `sdl2-2.0.0` dependency, making Halcyon
build it twice. First, `sdl2` is built along with all other dependencies
but then, after "Configuring app", Halcyon realises:

> Some add-source dependencies have been modified. They will be reinstalled...
> ...
> sdl2-2.0.0 (reinstall) changes: lens-4.9.1 added

Then `sdl2` is compiled again, and then our app, and in the end,
everything works. But it makes me still wonder if my configuration for
`sdl2` is right. Should it be mentioned in our `*.cabal` file or not?
And should it be in `.halcyon/constraints` or only
`.halcyon/sandbox-sources`?